Dispatch util functions into separate scripts
