+++ app/pages/Extractions_disponibles.py
@@ -9,13 +9,11 @@ from constants import (
 import pandas as pd
 from utils import (
     list_files,
-    disable_button,
     get_file_system,
     read_excel_from_s3,
-    display_pdf,
     format_extraction_name,
-    sidebar_content
 )
+from streamlit_utils import disable_button, display_pdf
 from pathlib import Path
 
 
@@ -23,7 +21,6 @@ st.set_page_config(layout="wide", page_title="Extractions disponibles", page_ico
 
 st.markdown("# Extractions disponibles")
 st.sidebar.header("Extractions disponibles")
-sidebar_content()
 st.write(
     """
     Liste des extractions disponibles.
